+++ src/input/visit.ts
@@ -183,10 +183,21 @@ function getIndexType(
   };
 }
 
-function normalizeTypeParameters(context: DeclVisitorContext, type: Type, decl: Type, types: Array<TypeModel>, refName: string) {
+function normalizeTypeParameters(
+  context: DeclVisitorContext,
+  type: Type,
+  decl: Type,
+  types: Array<TypeModel>,
+  refName: string,
+) {
   const maxTypes = (context.refs[refName] as any)?.types?.length ?? Number.MAX_SAFE_INTEGER;
   const typeParameterIds = decl?.typeParameters?.map(t => getDefaultTypeId(context, t)) ?? [];
-  const typeArgumentIds = type.aliasTypeArguments?.map(t => t.id) ?? [];
+  const typeArgumentIds = type.aliasTypeArguments?.map(t => t.id) ?? (type as any).typeArguments?.map(t => t.id) ?? [];
+
+  // omit types that are not listed on the original type parameters
+  while (types.length > maxTypes) {
+    types.pop();
+  }
 
   for (let i = typeParameterIds.length; i--; ) {
     const id = typeParameterIds[i];
@@ -195,11 +206,6 @@ function normalizeTypeParameters(context: DeclVisitorContext, type: Type, decl: 
       break;
     }
 
-    types.pop();
-  }
-
-  // omit types that are not listed on the original type parameters
-  while (types.length > maxTypes) {
     types.pop();
   }